requirements.txt
aioredis==1.3.1
amqp==2.6.0
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
avinit==1.2.2
billiard==3.6.3.0
cachetools==4.1.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-celery-beat==2.0.0
django-import-export==2.2.0
django-mysql==3.7.0
django-timezone-field==4.0
docker==4.2.2
docopt==0.4.0
et-xmlfile==1.0.1
future==0.18.2
geographiclib==1.50
geopy==1.22.0
google-api-core==1.22.0
google-auth==1.20.0
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gspread==3.6.0
hiredis==1.0.1
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.1
incremental==17.5.0
jdcal==1.4.1
jsonify==0.5
kombu==4.6.10
mailchimp==2.0.9
MarkupPy==1.14
msgpack==1.0.0
nodeenv==1.3.5
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.4
pandas==1.1.0
Pillow==7.1.2
protobuf==3.12.4
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygeocoder==1.2.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pypiwin32==223
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
websocket-client==0.57.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.1.0
